backend/requirements.txt
fastapi
uvicorn
langgraph
langchain-google-genai
python-dotenv
requests
python-multipart
pydantic
google-generativeai
sqlalchemy
aiosqlite
apscheduler
